+++ packages/saas-ui-date-picker/src/types.ts
@@ -1,10 +1,9 @@
-import {
-  DatePickerState,
-  DateRangePickerState,
-} from '@react-stately/datepicker'
+import { DatePickerState } from '@react-stately/datepicker'
+import { RangeValue } from '@react-types/shared'
 
 export type DateValue = DatePickerState['dateValue']
+export type TimeValue = DatePickerState['timeValue']
 export type FormattedValue = DateValue | Date | string | number
 
-export type DateRangeValue = DateRangePickerState['dateRange']
-export type TimeRangeValue = DateRangePickerState['timeRange']
+export type DateRangeValue = RangeValue<DateValue | null>
+export type TimeRangeValue = RangeValue<TimeValue | null>