+++ packages/saas-ui-menu/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @saas-ui/menu
 
+## 1.3.2
+
+### Patch Changes
+
+- @saas-ui/provider@1.1.8
+
 ## 1.3.1
 
 ### Patch Changes