+++ packages/saas-ui-modals/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @saas-ui/modals
 
+## 1.4.2
+
+### Patch Changes
+
+- @saas-ui/menu@1.3.2
+
 ## 1.4.1
 
 ### Patch Changes