+++ packages/saas-ui-provider/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @saas-ui/provider
 
+## 1.1.8
+
+### Patch Changes
+
+- Updated dependencies [69d328f]
+  - @saas-ui/theme@1.6.2
+
 ## 1.1.7
 
 ### Patch Changes